Modified server service to work with changes made in server control module
The name "kt" is replaced by "db" since this service is no more just for Kyoto Tycoon and should be able to run and maintain any database.
Instead of importing the functions, the class of KtServer is imported that contains all the functionalities in itself.
A function is added for getting the right server object (is going to be necessary after adding other databases)