Modified starting and stopping the database server
Any type of database can be started using DbServerService, just the dbElem has to be set for the desired type. (Currently it supports just Kyoto Tycoon)
For stopping the database, instead of calling stopKtserver(), the method of the server object, stopServer() is called. (This method has to be implemented for any future type of database)