Modified the redis server module to work with redis commands and options
'ktremotemgr' and 'ktserver' commands are replaced by 'redis-cli' and 'redis-server' respectively. The keywords that are being searched in the log file are updated to work with redis log file. The options for running the server are updated. There is no need to check the existence of two snapshots. Some TODOs are added.